+++ Cargo.toml
@@ -103,6 +103,10 @@ path = "examples/query.rs"
 name = "custom_renderer"
 path = "examples/custom_renderer.rs"
 
+[[example]]
+name = "disabled"
+path = "examples/disabled.rs"
+
 # Idiomatic Bevy code often triggers these lints, and the CI workflow treats them as errors.
 # In some cases they may still signal poor code quality however, so consider commenting out these lines.
 [lints.clippy]
